+++ R/weather.R
@@ -62,14 +62,15 @@
 #' @param string_as Character. What value to replace character strings in a
 #'   numeric measurement with. See Details.
 #' @param time_disp Character. Either "none" (default) or "UTC". See details.
-#' @param stn Data frame. The \code{stations} data frame to use. Will use the
-#'   one included in the package unless otherwise specified.
 #' @param encoding Character. Text encoding for download.
 #' @param list_col Logical. Return data as nested data set? Defaults to FALSE.
 #'   Only applies if `format = TRUE`
 #' @param verbose Logical. Include progress messages
 #' @param quiet Logical. Suppress all messages (including messages regarding
 #'   missing data, etc.)
+#' @param stn DEFUNCT. Now use `stations_dl()` to update internal data and
+#'   `stations_meta()` to check the date it was last updated.
+
 #'
 #' @return A tibble with station ID, name and weather data.
 #'
@@ -104,7 +105,7 @@ weather_dl <- function(station_ids,
                        format = TRUE,
                        string_as = NA,
                        time_disp = "none",
-                       stn = weathercan::stations,
+                       stn = NULL,
                        encoding = "UTF-8",
                        list_col = FALSE,
                        verbose = FALSE,
@@ -122,6 +123,13 @@ weather_dl <- function(station_ids,
   if(length(interval) > 1) {
     stop("'interval' must be either 'hour', 'day', OR 'month'")
   }
+
+  if(!is.null(stn)){
+    stop("`stn` is defunct, to use an updated stations data frame ",
+         "use `stations_dl()` to update the internal data, and ",
+         "`stations_meta()` to check when it was last updated", call. = FALSE)
+  }
+  stn <- stations()
 
   check_int(interval)