+++ data-raw/data-raw.R
@@ -1,6 +1,5 @@
 ## Get up-to-date stations data
-stations <- stations_dl()
-usethis::use_data(stations, overwrite = TRUE)
+stations_dl_internal(verbose = TRUE)
 
 kamloops_day <- weather_dl(51423, start = "2016-01-01", end = "2016-06-30", interval = "day")
 usethis::use_data(kamloops_day, overwrite = TRUE)