+++ NEWS.md
@@ -1,14 +1,18 @@
-# weathercan (development version)
+# weathercan 0.6.0
 
 ## Big changes
+- Move from data frame `stations` to function `stations()`. Returns same data
+  but is updateable with `stations_dl()` and you can check download dates
+  version with `stations_meta()`
 - Download climate normals from climate.weather.gc.ca
-  - More stations available
+  - More stations available (more than 2x as many!)
   - More year ranges available (1981-2010 and 1971-2000; 
     Note that while climate normals from 1961-1990 are available, they
     don't have climate ids making it tricky to download reliably)
     
 ## Small changes
 - Remove old deprecated function arguments
+- Add more test coverage
 
 ## Bug fixes
 - Download stations data frame from google drive rather than FTP site